% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fqa_citations.R
\docType{data}
\name{fqa_citations}
\alias{fqa_citations}
\title{Citations for Floristic Quality Assessment Databases}
\format{
A data frame
\describe{
\item{fqa_db}{Regional FQA database}
\item{recommended}{Indicates if the regional FQA database was recommended for use by the U.S. Army Corps of Engineers}
\item{notes}{Notes on the limitations or recommended usage of the regional FQA database}
\item{citation}{A citation for the regional FQA database}
...
}
}
\usage{
fqa_citations
}
\description{
A data set containing citations for the regional Floristic Quality Assessment (FQA)
databases contained in this package, as well as information on approval status,
and notes regarding the proper application of the regional database.
}
\keyword{datasets}
